# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TarotGen.io API",
        "description": "The backend for **TarotGen.io** — a PHP 8.5 / Slim 4 hybrid-REST API.\n\n- Resources are plural nouns (`/decks`, `/readings`); items are `/{id}`.\n- `GET` reads, `POST` creates (→`201`) or runs a command, `PUT` replaces,\n  `PATCH` applies a partial update / flag toggle, `DELETE` removes (→`204`).\n- State transitions that aren't a plain field-set are `POST …/{id}/{verb}`\n  (e.g. `…/finalize`, `…/approve`, `…/draw`).\n- Errors are `{ \"error\": \"…\" }`; multi-message validation is `{ \"errors\": […] }`.\n- Auth is a session cookie (`PHPSESSID`) set by the `/auth/*` endpoints —\n  there are no API keys or bearer tokens for first-party calls.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
...
        "/spreads": {
            "get": {
                "tags": [
                    "Spreads"
                ],
                "summary": "List all public spreads",
                "operationId": "5adec7eae6420940ee831f754e2f977d",
                "responses": {
                    "200": {
                        "description": "Array of spreads",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Spread"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Spreads"
                ],
                "summary": "Submit a spread into the moderation queue (rate-limited: 5/IP/hour)",
                "description": "Public endpoint: accept a user-submitted spread into the pending queue.\nRate-limited per IP to prevent spam. Never exposes the stored row.",
                "operationId": "c8dd5d463f67b61e59d160ecb62b0f24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "positions"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "positions": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Accepted into the pending queue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failure"
                    },
                    "429": {
                        "description": "Rate limit exceeded"
                    }
                }
            }
        },
        "/spreads/{spread_id}": {
            "get": {
                "tags": [
                    "Spreads"
                ],
                "summary": "A single public spread",
                "operationId": "c4565a7c707bdeb7725c417c3d679de9",
                "parameters": [
                    {
                        "name": "spread_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The spread",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Spread"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "InvalidSpreadID"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Spread": {
                "description": "A predefined tarot spread layout.",
                "properties": {
                    "spread_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "card_count": {
                        "type": "integer"
                    },
                    "positions": {
                        "description": "Array of position objects, each shaped:\n['order' => int, 'title' => string, 'x' => float, 'y' => float, 'rotation' => int]",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "description": "Session cookie issued on sign-in. Sent automatically by the browser.",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}